Some speed-up in cost computation. All costs end up as integers now, also target costs. Given that costs are of the order of 1 million, rounding errors can be neglected. Further speedup is needed.
git-svn-id: https://mary.opendfki.de/repos/trunk@120 953a6561-930b-0410-b2a6-db37d1b2ae63

+++ java/de/dfki/lt/mary/unitselection/viterbi/ViterbiCandidate.java
@@ -44,14 +44,14 @@ public class ViterbiCandidate {
 	private int pos = 0;
 	private Target target = null;
 	private ViterbiCandidate next = null;
-	private float targetCost = -1;
+	private int targetCost = -1;
 	
 	/**
 	 * Calculates and returns the target cost for this candidate
 	 * @param tcf the target cost function 
 	 * @return the target cost
 	 */
-	public float getTargetCost(TargetCostFunction tcf){
+	public int getTargetCost(TargetCostFunction tcf){
 	    if (targetCost == -1){
 	        targetCost = tcf.cost(target,unit);
 	        //System.out.println("Candidate with target "+target.getName()
